+++ docs/history.md
@@ -109,6 +109,12 @@ This is the first stable release of Hatch v1, a complete rewrite. Enjoy!
 ***Changed:***
 
 - When no build targets are specified on the command line, now default to `sdist` and `wheel` targets rather than what happens to be defined in config
+- The `code` version source now only supports files with known extensions
+
+***Added:***
+
+- The `code` version source now supports loading extension modules
+- Add `search-paths` option for the `code` version source
 
 ***Fixed:***